Поправил вывод текста в графическом режие. Перестал работать функционал PIC нужно через dbg дебажить

+++ src/kernel/interapts/interapt.h
@@ -7,7 +7,7 @@ void null_interapt(void);
 void init_video_interapt(void);
 void clear_screen_video_interapt(void);
 void set_cursor_video_interapt(uint32_t x, uint32_t y);
-void set_tem_terminal_interapt(uint8_t background, uint8_t text_color);
+void set_tem_terminal_interapt(uint16_t background, uint16_t text_color);
 void print_char_interapt(uint8_t sumbole);
 
 #endif
